fix: #27 エンティティ定義位置を修正
